requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
httpx==0.28.1
python-dotenv==1.0.1
